requirements.txt
streamlit
tensorflow
numpy
pillow
opencv-python-headless
streamlit_drawable_canvas
rembg